refactor: move events to `events` entities submodule
