+++ src/client/websocket.rs
@@ -8,7 +8,7 @@ use {
 };
 
 use crate::{
-    entities::{ClientToServerEvent, ServerToClientEvent},
+    entities::events::{ClientToServerEvent, ServerToClientEvent},
     error::Error,
     Result,
 };
