+++ src/entities/events/channel/delete.rs
@@ -1,4 +1,4 @@
-use crate::entities::ServerToClientEvent;
+use super::super::ServerToClientEvent;
 
 /// A channel has been deleted.
 #[derive(Debug)]
