+++ src/entities/events/channel/stop_typing.rs
@@ -1,4 +1,4 @@
-use crate::entities::ServerToClientEvent;
+use super::super::ServerToClientEvent;
 
 /// A user has stopped typing in a channel.
 #[derive(Debug)]
