+++ src/entities/events/server/member_join.rs
@@ -1,4 +1,4 @@
-use crate::entities::ServerToClientEvent;
+use super::super::ServerToClientEvent;
 
 /// A user has joined the group.
 #[derive(Debug)]
