+++ src/entities/events/server/member_leave.rs
@@ -1,4 +1,4 @@
-use crate::entities::ServerToClientEvent;
+use super::super::ServerToClientEvent;
 
 /// A user has left the group.
 #[derive(Debug)]
