+++ src/entities/events/server/role_update.rs
@@ -1,6 +1,6 @@
 use {serde::Deserialize, serde_json::Value as Json};
 
-use crate::entities::ServerToClientEvent;
+use super::super::ServerToClientEvent;
 
 /// Specifies a field to remove on server role update.
 #[derive(Debug, Deserialize, Clone, Copy, PartialEq)]
